Another refactor to pull mysql under sql package
