+++ testonly/integration/clientserver.go
@@ -31,7 +31,7 @@ import (
 	"github.com/google/trillian/extension"
 	"github.com/google/trillian/server"
 	"github.com/google/trillian/storage/sql/coresql"
-	"github.com/google/trillian/storage/mysql"
+	"github.com/google/trillian/storage/sql/mysql"
 	stestonly "github.com/google/trillian/storage/testonly"
 	"github.com/google/trillian/testonly"
 	"github.com/google/trillian/util"
